Refactor: 바텀시트 공용 훅 사용하도록 수정
