+++ src/app/[userNickname]/[listId]/_components/ListDetailOuter/ModalButtonStyle.css.ts
@@ -4,6 +4,6 @@ export const resetButtonStyle = style({
   background: 'none',
 });
 
-export const deleteButton = style({
+export const buttonCursor = style({
   cursor: 'pointer',
 });
